+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog for languages
+## 2015-06-05
+  * Added 2 new entries for server connection status.
 
 ## 2015-05-30
   * Added handling of word count for the editor.